requirements.txt
# F5-TTS Vast.ai Automation Requirements
# Install with: pip install -r requirements.txt

# PyTorch (CUDA 11.8 compatible)
torch>=2.0.0
torchaudio>=2.0.0

# Whisper AI for transcription
openai-whisper>=20231117

# Google Drive integration
pydrive2>=1.15.0

# Audio processing
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional: For better audio quality
resampy>=0.4.0
audioread>=3.0.0
